feat: metrics draft

+++ metrics.go
@@ -3,8 +3,8 @@ package noise
 // metrics hold the statistics related to remote peers.
 // We can add any method related to adaptive lookup logic here.
 type metrics struct {
-	handshakeStart uint64
-	handshakeEnd   uint64
+	handshakeStart uint32
+	handshakeEnd   uint32
 	latency        uint16
 	bandwidth      uint16
 }
